Create Projection Noise
Here we are calculating mean and standard deviation of frequency of observing |0> state over many trials. This is representation of Projection Noise that is introduced due to the finite number of times that we repeat the experiment. After plotting the graphs it is seen that the variance and standard deviation decrease as the number of shots are increased.